moving mockinterface into rosinterface.mock since design follows ROS concepts. fixed all tests. bumped pyros minor version to 0.3.0 because of all the changes...
