+++ CMakeLists.txt
@@ -27,7 +27,7 @@ if (CATKIN_ENABLE_TESTING)
     ##############################################################################
 
     # Careful this will not support parallel testing, due to multiprocess environment used in tests
-    catkin_add_nosetests(pyros/mockinterface/tests)
+    catkin_add_nosetests(pyros/rosinterface/mock/tests)
     # because of ROS (node_init() can be called only once per process)
     # these should be run in separate process each to avoid global side effect during tests
     # TODO : find a way to force nose to run each test isolated ?
