+++ pyros/rosinterface/mock/mockmessage.py
