+++ pyros/rosinterface/mock/mockparam.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, print_function
 
-from ..baseinterface import TransientIf
+from ...baseinterface import TransientIf
 
 
 class MockParam(TransientIf):
