+++ pyros/rosinterface/mock/tests/test_mockservice.py
@@ -4,8 +4,8 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..')))
 
-from pyros.mockinterface import extract_values, populate_instance, FieldTypeMismatchException, NonexistentFieldException, StatusMsg
-from pyros.mockinterface import statusecho_service, MockService
+from pyros.rosinterface.mock import extract_values, populate_instance, FieldTypeMismatchException, NonexistentFieldException, StatusMsg
+from pyros.rosinterface.mock import statusecho_service, MockService
 from nose.tools import assert_true, assert_false, assert_raises
