+++ setup.py
@@ -88,9 +88,9 @@ setuptools.setup(name='pyros',
         'pyros',
         'pyros.tests',
         'pyros.baseinterface',
-        'pyros.mockinterface',
-        'pyros.mockinterface.tests',
         'pyros.rosinterface',
+        'pyros.rosinterface.mock',
+        'pyros.rosinterface.mock.tests',
         'pyros.rosinterface.tests',
         'pyros.rosinterface.rostests',
         'pyros.zmpinterface',
